requirements.txt
psycopg2-binary==2.9.9
pygame==2.5.2
pandas==2.1.3 